refactor: add pages folder in file structure
