+++ src/components/App.spec.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import {render} from '@testing-library/react';
 import App from './App';
 
-jest.mock('./PageMap', () => ({PageMap: () => <div>PageMap component</div>}));
-jest.mock('./PageProfile', () => ({PageProfile: () => <div>PageProfile component</div>}));
-jest.mock('./PageLogin', () => ({PageLogin: () => <div>PageLogin component</div>}));
+jest.mock('./pages/PageMap', () => ({PageMap: () => <div>PageMap component</div>}));
+jest.mock('./pages/PageProfile', () => ({PageProfile: () => <div>PageProfile component</div>}));
+jest.mock('./pages/PageLogin', () => ({PageLogin: () => <div>PageLogin component</div>}));
 
 describe('App', () => {
   it('renders correctly', () => {
