+++ src/components/pages/PageMap.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import {Map} from './Map';
-import {MapForm} from './MapForm';
-import ErrorBoundary from './ErrorBoundary';
-import {withAuth} from './hoc/AuthContext';
+import {Map} from '../Map';
+import {MapForm} from '../MapForm';
+import ErrorBoundary from '../ErrorBoundary';
+import {withAuth} from '../hoc/AuthContext';
 
 export const PageMap = () => (
   <ErrorBoundary>
