+++ src/components/pages/PageRegistration.jsx
